remove macros from bytes module
