+++ src/bytes/mod.rs
@@ -1,6 +1,6 @@
 //! Parsers recognizing bytes streams
 
-#[macro_use]
-mod macros;
 pub mod complete;
 pub mod streaming;
+#[cfg(test)]
+mod tests;
